+++ src/components/CommentCreator.vue
@@ -1,6 +1,6 @@
 <template>
   <form @submit="handleSubmit" :class="{ waiting: isWaiting }">
-    <MarkdownEditorWrapper ref="markdownEditor" :disabled="isWaiting" :hideFileUpload="true" />
+    <MarkdownEditorWrapper ref="markdownEditor" :disabled="isWaiting" :id="parentIdNotNull" />
     <input v-if="!isWaiting" class="form-submit" type="submit" value="Opret kommentar!">
     <span v-else class="waiting-text">Vent venligst..</span>
     <div v-if="isWaiting" class="overlay"></div>
@@ -17,8 +17,17 @@ export default {
   },
 
   data() {
+    // Generate unique file list ID
+    const count = this.$store.state.counter;
+    this.$store.dispatch('counter');
+    const curPost = this.$store.state.forum.currentPost.id;
+    const parentIdNotNull = this.parentId === null
+      ? `${curPost}-rootComment-${count}`
+      : this.parentId;
+
     return {
       isWaiting: false,
+      parentIdNotNull,
     };
   },
 
@@ -34,6 +43,7 @@ export default {
       event.preventDefault();
 
       const commentContent = this.$refs.markdownEditor.getValue();
+      const files = this.$refs.markdownEditor.getFiles();
 
       if (commentContent.trim().length === 0) {
         this.$dialog.alert('Indholdet må ikke være tomt!');
@@ -50,6 +60,7 @@ export default {
           forumPathName: forum,
           body: commentContent,
           parentId: this.parentId,
+          files,
         });
 
         this.$emit('comment-created');