+++ src/store/actions.js
@@ -115,4 +115,28 @@ export default {
     const { authToken } = context.rootState.auth;
     await api.updateAvatar(authToken, newAvatar);
   },
+
+  toggleSuperuser(context, userId) {
+    const { authToken } = context.rootState.auth;
+    return api.toggleSuperuser(authToken, userId);
+  },
+
+  toggleBan(context, userId) {
+    const { authToken } = context.rootState.auth;
+    return api.toggleBan(authToken, userId);
+  },
+
+  async deleteUser(context, userId) {
+    const { authToken } = context.rootState.auth;
+    await api.deleteUser(authToken, userId);
+
+    const updatedUsersList = context.state.allUsers.filter(user => user.id !== userId);
+    context.commit('SET_USERS', updatedUsersList);
+  },
+
+  counter(context) {
+    const val = context.state.counter;
+    context.commit('INCREMENT_COUNTER');
+    return val;
+  },
 };